requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
line-bot-sdk